harbor-platform/cognee-service/pyproject.toml
# Cognee Microservice for HARBOR
# Graph RAG service using Cognee

[project]
name = "harbor-cognee-service"
version = "0.1.0"
description = "Graph RAG microservice for HARBOR platform using Cognee"
requires-python = ">=3.10"
dependencies = [
    "cognee>=0.1.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
